522: v1.9: add checks
